Bugfixes and new feature
Fixed error where uncertainty scenarios DVHs were exported wrongly for min, max og mean dose.
Fixed error where the dose for OAR_3 was not read and used.
Changed the plan names not to start with a number.
Fixed the limit for the difference in MU to 1% and added a check of the dose normalization of photon plans, as this was sometimes wrong.
Added the ability to select a sum plan for evaluation. This required a restructuring of some methods throughout the script.