Changes to DVH export format, small bugfix
Adjustment to the DVH export where D1cc, D0.5 cc and D0.05 cc are read and exported in the file. The default Gy step size is set to 0.05 Gy and the rounding of all numbers in the whole script is now mathematically correct.
A small fix where the selected plan in Eclipse was used to read of some values that then leads to a crash was corrected.